server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pgvector==0.3.6
python-dotenv==1.0.1
anthropic==0.40.0
sentence-transformers==3.3.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
httpx==0.28.1
redis==5.2.1
celery==5.4.0